Modified XML_Reader interface with new ENUM values for information control in and out Modified Login Reader Error Messages for Clarity, and Values checks for more NULLS to remove the change of errors
Modified Account Reader Error Messages for Clarity, Logic Statements for Readability, and Values checks for more NULLS to remove the change of errors

Modified Medical Reader Code due to errors from previous tests for Account and Login. This resulted in cleaner code and less methods as some where discovered to be nested for no reason. Modified the layout of the XML expected as well to make it easier to read and catch errors
 Modified Error Messages for Clarity, Logic Statements for Readability, and Values checks for more NULLS to remove the change of errors

 Added Test Files for Medical reader, including new XML files